fix translate start
