+++ app/src/main/java/com/sopt/smeme/business/viewmodel/Step2ViewModel.kt
@@ -1,58 +1,45 @@
 package com.sopt.smeme.business.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.sopt.smeme.BuildConfig
-import com.sopt.smeme.bridge.controller.PapagoAPI
-import com.sopt.smeme.bridge.controller.TranslateResult
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-import javax.inject.Inject
 
-@HiltViewModel
-class Step2ViewModel @Inject constructor(
-    private val papagoAPI: PapagoAPI
-) : ViewModel() {
-    fun translate(
-        text: String,
-        onCompleted: (String) -> Unit = {},
-        onError: (Throwable) -> Unit = {}
-    ) {
-        viewModelScope.launch {
-            // suspend 호출
-            println("######################################## first")
-
-            papagoAPI.transferPapago(
-                "ko", "en", text,
-                BuildConfig.CLIENT_ID, BuildConfig.CLIENT_SECRET
-            ).enqueue(object : Callback<TranslateResult> {
-                override fun onResponse(
-                    call: Call<TranslateResult>,
-                    response: Response<TranslateResult>
-                ) {
-                    if (response.isSuccessful) {
-                        val result = response.body()?.message?.result?.translatedText
-
-                        if (result != null) {
-                            onCompleted.invoke(result)
-                            Log.d("번역 성공", "변역된 일기:$result")
-                        } else {
-                            throw IllegalStateException("TODO : change to SmemeException")
-                        }
-                    } else {
-                        onError.invoke(IllegalStateException("번역이 정상적으로 수행되지 않았습니다. ${response.code()} : ${response.message()}"))
-                    }
-                }
-
-                override fun onFailure(call: Call<TranslateResult>, t: Throwable) {
-                    onError.invoke(t)
-                }
-            })
-        }
-    }
-
+//@HiltViewModel
+class Step2ViewModel : ViewModel() {
+//    private val papagoAPI: PapagoAPI
+//) : ViewModel() {
+//    fun translate(
+//        text: String,
+//        onCompleted: (String) -> Unit = {},
+//        onError: (Throwable) -> Unit = {}
+//    ) {
+//        viewModelScope.launch(Dispatchers.IO) {
+//            // suspend 호출
+//            papagoAPI.transferPapago(
+//                "ko", "en", text,
+//                BuildConfig.CLIENT_ID, BuildConfig.CLIENT_SECRET
+//            ).enqueue(object : Callback<TranslateResult> {
+//                override fun onResponse(
+//                    call: Call<TranslateResult>,
+//                    response: Response<TranslateResult>
+//                ) {
+//                    if (response.isSuccessful) {
+//                        var result = response.body()?.message?.result?.translatedText
+//
+//                        if (result != null) {
+//                            onCompleted.invoke(result)
+//                            Log.d("번역 성공", "변역된 일기:$result")
+//                        } else {
+//                            throw IllegalStateException("TODO : change to SmemeException")
+//                        }
+//                    } else {
+//                        onError.invoke(IllegalStateException("번역이 정상적으로 수행되지 않았습니다. ${response.code()} : ${response.message()}"))
+//                    }
+//                }
+//
+//                override fun onFailure(call: Call<TranslateResult>, t: Throwable) {
+//                    onError.invoke(t)
+//                }
+//            })
+//        }
+//    }
+//
 }
